+++ python_modules/dagster/dagster/core/errors.py
@@ -18,66 +18,6 @@ class DagsterUserError(DagsterError):
     pass
 
 
-class DagsterEvaluationErrorReason(Enum):
-    RUNTIME_TYPE_MISMATCH = 'RUNTIME_TYPE_MISMATCH'
-    MISSING_REQUIRED_FIELD = 'MISSING_REQUIRED_FIELD'
-    FIELD_NOT_DEFINED = 'FIELD_NOT_DEFINED'
-    SELECTOR_FIELD_ERROR = 'MULTIPLE_FIELDS_DEFINED'
-
-
-FieldNotDefinedErrorData = namedtuple('FieldNotDefinedErrorData', 'field_name')
-MissingFieldErrorData = namedtuple('MissingFieldErrorData', 'field_name')
-RuntimeMismatchErrorData = namedtuple('RuntimeMismatchErrorData', 'dagster_type value_rep')
-
-ERROR_DATA_TYPES = (
-    FieldNotDefinedErrorData,
-    MissingFieldErrorData,
-    RuntimeMismatchErrorData,
-)
-
-
-class EvaluationStack(namedtuple('_EvaluationStack', 'entries')):
-    def __new__(cls, entries):
-        return super(EvaluationStack, cls).__new__(
-            cls,
-            check.list_param(entries, 'entries', of_type=EvaluationStackEntry),
-        )
-
-    @property
-    def levels(self):
-        return [entry.field_name for entry in self.entries]
-
-
-class EvaluationStackEntry(namedtuple('_EvaluationStackEntry', 'field_name field_def')):
-    def __new__(cls, field_name, field_def):
-        # TODO: fix circular import
-        from dagster.core.types import Field
-        return super(EvaluationStackEntry, cls).__new__(
-            cls,
-            check.str_param(field_name, 'field_name'),
-            check.inst_param(field_def, 'field_def', Field),
-        )
-
-
-class EvaluationError(namedtuple('_EvaluationError', 'stack reason message error_data')):
-    def __new__(cls, stack, reason, message, error_data):
-        return super(EvaluationError, cls).__new__(
-            cls,
-            check.inst_param(stack, 'stack', EvaluationStack),
-            check.inst_param(reason, 'reason', DagsterEvaluationErrorReason),
-            check.str_param(message, 'message'),
-            check.opt_inst_param(error_data, 'error_data', ERROR_DATA_TYPES),
-        )
-
-
-class DagsterEvaluateConfigValueError(DagsterError):
-    '''Indicates invalid value was passed to a type's evaluate_value method'''
-
-    def __init__(self, stack, *args, **kwargs):
-        super(DagsterEvaluateConfigValueError, self).__init__(*args, **kwargs)
-        self.stack = check.inst_param(stack, 'stack', EvaluationStack)
-
-
 class DagsterRuntimeCoercionError(DagsterError):
     '''Runtime checked faild'''
     pass
